modificación en media de peso

+++ vaterinaria/src/main/java/co/edu/uniquindio/poo/Mascota.java
@@ -3,7 +3,7 @@
 
 package co.edu.uniquindio.poo;
 
-public record Mascota(String nombre, String especie, String raza, byte edad, String género, String  color, byte peso ) {
+public record Mascota(String nombre, String especie, String raza, byte edad, String género, String  color, Double peso ) {
     public Mascota{
     }
 }
